REVERTME [VPG]: drm/i915/vlv: Support fb width and height less than 3 in plane updates.
Add support in plane update to accept fb width and height < 3 for VLV
based devices.
User layer can send fb width and height < 3 which are valid
for plane updates in some instances and we do not want to
discard the update.

Patch is currently REVERTME as it is being reviewed if the limits can be relaxed for all platforms.

Change-Id: I5c9d8d30730cbb64484d85722662a55ff3bad298
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-16129
Signed-off-by: Vidya Srinivas <vidya.srinivas@intel.com>
Reviewed-on: https://android.intel.com:443/316844