Add Marvin condition for ISP and camera related code
Wake up of PCI device 01.0 takes aproximately 7 seconds because ISP is
disabled in firmware on Marvin and ATOMISP driver is deactivated. With
such conditions the call to pmu_nc_set_power_state is triggered and
pass the wrong ISP power island address which fires a timeout. Besides,
on Marvin we don't need to power down ISP here.

Variable camera_off has to be set to true all the time on Marvin since
ISP is disabled. It causes trouble with function nc_device_state()
returning an incorrect status.

Change-Id: If64ab5f1463bff975befac76d6689c0d7581b30e
Tracked-On: https://jira01.devtools.intel.com/browse/MARVIN-120
Signed-off-by: Etienne Lachaud <etiennex.lachaud@intel.com>
Reviewed-on: https://android.intel.com:443/317097