[Telephony]: memory leak in drv pdata MCD
When the driver is removed from the system, the
allocated memory to store the ACPI data, recovered
during the driver probe are not freed.
Moreover, the object used to parse the ACPI table is
not freed after used.
This patch solves this issue by adding the
corresponding kfree calls for the corresponding
memory allocations.

Change-Id: Iac721d9b5cc6c72a4f49d7cde0368d61e79c8026
Tracked-On: https://jira01.devtools.intel.com/browse/IMINAN-702
Signed-off-by: Radu Stancu <radu.georgex.stancu@intel.com>
Reviewed-on: https://android.intel.com:443/318587