Vibra: move board specific initialization to probe function
Android's vibrator detection is done via a sysfs entry.
Android uses those sysfs entries to communicate with the driver.
The sysfs entries are created in the init() function, which is
executed on all platforms.
On Saltbay, the driver is not probed but still detected by the
Android system due to the sysfs entry.

This patch moves the device specific sysfs initialization to the
probe function, so that it is only executed on compatible devices.

Change-Id: Id0eed881024ba532aabefe8cec0582f7efa7f6e8
Tracked-On: https://jira01.devtools.intel.com/browse/MARVIN-385
Signed-off-by: Guillaume Marcenac <guillaumex.marcenac@intel.com>
Reviewed-on: https://android.intel.com:443/340422