display: full init code for tianma panel
The MTP settings for early Tianma panels are incorrect.
This requires the full initcode to be sent from the driver.

The new command line parameter 'panel_init_code=1/0' selects whether:
- 0: send the minimal settings for panels with the correct MTP settings
- 1: send the full MTP settings
If parameter is not on the commandline, default is 0.

Also add readback of all panel registers if TIANMA_DEBUG is defined.

Change-Id: Iacf94ca90fc7b530433757f049d8f8ee12320d90
Tracked-On: https://jira01.devtools.intel.com/browse/MARVIN-542
Signed-off-by: Sophia Gong <sophia.gong@intel.com>
Signed-off-by: Saadi Maalem <saadi.maalem@intel.com>
Reviewed-on: https://android.intel.com:443/353033