Haptics: add sysfs entry on vibrator driver
JIRA: MARVIN-145

Add calibration status, OTP status.
Add sysfs to run the calibration.
Add sysfs to run the save on otp memory.
Be careful, saving on otp memory can be done only one time.
To be able to create this sysfs we have to change the sequence of
the probe (we define drv260x->client earlier)

We need this sysfs access to be able to save the calibration.
It will be used during the factory.

Usage:
"cat calibration" return the status(never launch, ongoing, calibrated)
"cat OTP_status" return the status(memory lock or unlock)
"echo 1 > do_calibration" run the calibration sequence
"cat do_calibration" tell if the calibration is still ongoing
"echo 1 > OTP_save" save the value in the otp memory

Change-Id: If555be5ebf7fb38a8200b9880f86ecd568d6c975
Signed-off-by: Guillaume Marcenac <guillaumex.marcenac@intel.com>
Reviewed-on: https://android.intel.com/373190
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Pujol, Benjamin <benjamin.pujol@intel.com>