charger: workaround to disable charger IC
JIRA: MARVIN-347

The charger IC on Marvin is not connected to the PMIC through I2C.
Due to the NVM configuration, it is not possible to disable the charger.

To workaround this issue, until a fix is done at hardware level, the SWCONTROL
bit inside the PMIC is toggled to enable/disable charging.
This will work for charging over wireless or SDP, but not over DCP.

Change-Id: I841cd469d8594c44d800345f775f3bbb130d4f3c
Signed-off-by: Saadi Maalem <saadi.maalem@intel.com>
Reviewed-on: https://android.intel.com/385582
Reviewed-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>
Tested-by: Tasayco Loarte, VictorX <victorx.tasayco.loarte@intel.com>