wlc: implement docking detection mechanism
Wireless Charger can be stopped during the charging stage to cool down the
battery or for trickle charging.
The driver should be able to detect if the device has been moved or if it is
still on the charger.
This patch relies on a callback to receive the undocking event.
The callback is registered through the function pointer
wc_docking_detection which is defined as part of the platform data.
Docking state is reported to the system through the 'online' power_supply entry.

Change-Id: Ia3215c14fdfdbe16095d00fac7b2705c9b825fbe
Tracked-On: https://jira.ndg.intel.com/browse/MARVIN-444
Signed-off-by: Patrick Combes <patrick.combes@intel.com>
Signed-off-by: Honore Tricot <honorex.tricot@intel.com>
Reviewed-on: https://android.intel.com/395238
Reviewed-by: jenkins_ndg <jenkins_ndg@intel.com>
Reviewed-by: Maalem, Saadi <saadi.maalem@intel.com>